+++ Campus.Application/Attendances/Queries/GetAllLectorsAttendances/GetAllLectorsAttendancesQueryHandler.cs
@@ -33,7 +33,7 @@ namespace Campus.Application.Attendances.Queries.GetAllLectorsAttendances
                 .Select(x => new ChartDataItem
                 {
                     Date = x.Date,
-                    AttendancePercentage = (x.StudentsCount / x.Lesson.Group.StudentsCount) * 100
+                    AttendancePercentage = ((double)x.StudentsCount / x.Lesson.Group.StudentsCount) * 100
                 }).ToListAsync();
 
             return chartData;
